requirements.txt
Python = 3.7.0
Django = 2.0.7
djangorestframework = 3.9.1
django-filter = 2.1.0
